requirements.txt
opencv-python
numpy
pyyaml
mediapipe
ultralytics
